Added main html, css, and js files
Also added TheroHTMLtest(1).html, which is the html doc generated from the xslt that Dan and I built. The learning goal should be to have the xslt build out the html page directly into its final form; the likely goal will be to end up with a finished html product, irrespective of the xml/xslt upstream of it.